refactor: remove duplicate useKeyboardLock.ts file
